// src/typings/react/DevToolsHookTypes.ts
export type Handler = (data: any) => void;

export type NativeType = Object;

export type RendererID = number;

export type ResolveNativeStyle = (stylesheetID: any) => Object | null | undefined;
// ...
export type ReactRenderer = {
  findFiberByHostInstance: (hostInstance: NativeType) => Fiber | null | undefined;
  version: string;
  bundleType: BundleType;

  // 16.9+
  overrideHookState?: (fiber: Object, id: number, path: Array<string | number>, value: any) => void | null | undefined;

  // 16.7+
  overrideProps?: (fiber: Object, path: Array<string | number>, value: any) => void | null | undefined;

  // 16.9+
  scheduleUpdate?: (fiber: Object) => void | null | undefined;
  setSuspenseHandler?: (shouldSuspend: (fiber: Object) => boolean) => void | null | undefined;

  // Only injected by React v16.8+ in order to support hooks inspection.
  currentDispatcherRef?: { current: null | Dispatcher };

  // Only injected by React v16.9+ in DEV mode.
  // Enables DevTools to append owners-only component stack to error messages.
  getCurrentFiber?: () => Fiber | null;

  // Uniquely identifies React DOM v15.
  ComponentTree?: any;

  // Present for React DOM v12 (possibly earlier) through v15.
  Mount?: any;
};

export type RendererInterface = {
  cleanup: () => void;
  findNativeNodesForFiberID: FindNativeNodesForFiberID;
  flushInitialOperations: () => void;
  getBestMatchForTrackedPath: () => PathMatch | null;
  getFiberIDForNative: GetFiberIDForNative;
  getInstanceAndStyle(id: number): InstanceAndStyle;
  getProfilingData(): ProfilingDataBackend;
  getOwnersList: (id: number) => Array<Owner> | null;
  getPathForElement: (id: number) => Array<PathFrame> | null;
  handleCommitFiberRoot: (fiber: Object, commitPriority?: number) => void;
  handleCommitFiberUnmount: (fiber: Object) => void;
  inspectElement: (id: number, path?: Array<string | number>) => InspectedElementPayload;
  logElementToConsole: (id: number) => void;
  overrideSuspense: (id: number, forceFallback: boolean) => void;
  prepareViewElementSource: (id: number) => void;
  renderer: ReactRenderer | null;
  setInContext: (id: number, path: Array<string | number>, value: any) => void;
  setInHook: (id: number, index: number, path: Array<string | number>, value: any) => void;
  setInProps: (id: number, path: Array<string | number>, value: any) => void;
  setInState: (id: number, path: Array<string | number>, value: any) => void;
  setTraceUpdatesEnabled: (enabled: boolean) => void;
  setTrackedPath: (path: Array<PathFrame> | null) => void;
  startProfiling: (recordChangeDescriptions: boolean) => void;
  stopProfiling: () => void;
  updateComponentFilters: (componentFilters: Array<ComponentFilter>) => void;
};

export type DevToolsHook = {
  listeners: {
    [key: string]: Array<Handler>;
  };
  rendererInterfaces: Map<RendererID, RendererInterface>;
  renderers: Map<RendererID, ReactRenderer>;

  emit: (event: string, data: any) => void;
  getFiberRoots: (rendererID: RendererID) => Set<Object>;
  inject: (renderer: ReactRenderer) => number | null;
  on: (event: string, handler: Handler) => void;
  off: (event: string, handler: Handler) => void;
  reactDevtoolsAgent?: Object | null | undefined;
  sub: (event: string, handler: Handler) => () => void;

  // Used by react-native-web and Flipper/Inspector
  resolveRNStyle?: ResolveNativeStyle;
  nativeStyleEditorValidAttributes?: ReadonlyArray<string>;

  // React uses these methods.
  checkDCE: (fn: Function) => void;
  onCommitFiberUnmount: (rendererID: RendererID, fiber: Fiber) => void;
  onCommitFiberRoot: (rendererID: RendererID, root: FiberRoot, commitPriority?: number) => void;
};

export type DevToolsHookSorge = Pick<DevToolsHook, 'inject' | 'onCommitFiberUnmount' | 'onCommitFiberRoot'> & {
    // This is a legacy flag.
    // React v16 checks the hook for this to ensure DevTools is new enough.
    supportsFiber: boolean;
  } & Partial<DevToolsHook>;